% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_indicators.R
\name{predict_indicators}
\alias{predict_indicators}
\title{Predict ecological quality group using indicator taxa}
\usage{
predict_indicators(abundance_matrix, indicator_table, true_group = NULL)
}
\arguments{
\item{abundance_matrix}{Matrix/data.frame of taxa abundances (samples x taxa)}

\item{indicator_table}{Output from run_indval_analysis()}

\item{true_group}{Optional vector with true ecological group for each sample}
}
\value{
A list with predictions, optional accuracy and confusion matrix
}
\description{
Predict ecological quality group using indicator taxa
}
