% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_AQI.R
\name{calculate_AQI}
\alias{calculate_AQI}
\title{Calculate Aquatic Quality Index (AQI)}
\usage{
calculate_AQI(data, thresholds)
}
\arguments{
\item{data}{Data frame containing Site column and contaminant concentrations.}

\item{thresholds}{Data frame with columns: contaminant and limit.}
}
\value{
Data frame with F1, F2, F3 and final AQI per site.
}
\description{
Calculates AQI following the CCME-WQI conceptual framework
using contaminant concentration thresholds (e.g., PNEC values).
}
