% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_WQI.R
\name{calculate_WQI}
\alias{calculate_WQI}
\title{Calculate Water Quality Index (WQI)}
\usage{
calculate_WQI(data, limits)
}
\arguments{
\item{data}{Data frame containing physicochemical parameters}

\item{limits}{Named list of threshold functions}
}
\value{
Data frame with WQI values
}
\description{
Calculate Water Quality Index (WQI)
}
