% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_indval_analysis.R
\name{run_indval_analysis}
\alias{run_indval_analysis}
\title{Run Indicator Species Analysis}
\usage{
run_indval_analysis(
  asv_matrix,
  groups,
  method = "IndVal.g",
  nperm = 999,
  presence_absence = TRUE
)
}
\arguments{
\item{asv_matrix}{Matrix or data frame (sites × taxa).}

\item{groups}{Vector of group assignments per site.}

\item{method}{Association method ("IndVal.g", "r", "r.g").}

\item{nperm}{Number of permutations.}

\item{presence_absence}{Logical; convert abundances to PA.}
}
\value{
multipatt object.
}
\description{
Performs indicator species analysis using IndVal, rphi or r.g
from the indicspecies package.
}
